+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 * Support for displaying speed at which the CPU is running in web debug
+* Headless execution mode in Boytacean SDL
+* Many more parameters added for Boytacean SDL
 
 ### Changed